rework headers into seperated core_lib.h and core_lib_internal.h
